jokes_app_backend/pyproject.toml
[tool.poetry]
name = "jokes-app-backend"
version = "0.1.0"
description = "training app backend"
authors = ["jsgerardor <gerardo.rios@ioet.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
flake8 = "^7.0.0"
autoflake = "^2.2.1"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
faker = "^22.1.0"
pytest-asyncio = "^0.23.3"
ipython = "^8.20.0"
anyio = "^4.2.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"